split sarifutils and move test funcs
